requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.26.5
gunicorn==21.2.0